+++ module/Shop/Service/SiteService.php
@@ -52,6 +52,17 @@ final class SiteService implements SiteServiceInterface
 	}
 
 	/**
+	 * Returns minimal product's price associated with provided category id
+	 * 
+	 * @param string $categoryId
+	 * @return float
+	 */
+	public function getMinCategoryPriceCount($categoryId)
+	{
+		return $this->productManager->getMinCategoryPriceCount($categoryId);
+	}
+
+	/**
 	 * Returns an array of entities with products that have maximal view count
 	 * 
 	 * @param integer $limit