+++ module/Shop/CHANGELOG.md
@@ -4,8 +4,9 @@ CHANGELOG
 1.2
 ---
 
+ * Added support for starting prices in category. This functionality is implemented in `SiteService::getMinCategoryPriceCount`
  * Implemented support for "Mostly Viewed Products". Added `getProductsWithMaxViewCount()` to `SiteService` 
- * Added support views count
+ * Added support for views count
  * Minor improvements in code base
  * Added support for permanent URLs. Now each product's entity has `getPermanentUrl()` method
  * Added `SiteService`, which is called `$shop`. Moved all related functionality there as well